+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Note that since we don't clearly distinguish between a public and private interf
     - Artistic option fits spherical harmonics to bins
 - Fix camera reset handling for (temporary) empty scenes (#1903)
 - Remove `firstStepSize` tracing parameter, derive automatically
+- Fix illumination `auto` thickness mode never being applied
 
 - Camera improvements
   - Support multiple camera transition shapes